chore: move dataset import to utils
