+++ libs/core/kiln_ai/utils/test_dataset_import.py
@@ -9,13 +9,13 @@ from pathlib import Path
 
 import pytest
 
-from kiln_ai.adapters.dataset_import import (
+from kiln_ai.datamodel import Project, Task
+from kiln_ai.utils.dataset_import import (
     DatasetFileImporter,
     DatasetImportFormat,
     ImportConfig,
     KilnInvalidImportFormat,
 )
-from kiln_ai.datamodel import Project, Task
 
 logger = logging.getLogger(__name__)
